GPSequence mode now in Arbotix Controller +++
I now have the GPSequence mode in the Arbotix controller, R1 cycles
through the sequences, may still update this one as well as PS2 not to
simply count to 5, but instead test to see if a sequence is defined.

Also other debug commands available in background terminal, but probably
only available with DIY XBee  or PS2 controller.

Remember - WIP