Running Sequence when robot is upside down
Have code in place for sequences stored in PROGMEM that inverts it when
the robot is upside down.  Need to extend to EEPROM.  Also I hacked up
my own version of BioloidEx to have a setPose that allows me to pass in
an index instead of ID as to keep from searching every ... time.   Will
probably update other code to use this