Teensy merge, Auto XBee, ...
Some of the stuff I did for the Serial version was merged here, plus
other stuff for Teensy 3.1.  In addition, put in code for detecting XBee
that has not been configured and configuring it.  Also found issue where
sometimes XBee would init into command mode at 9600 and not respond for
10 seconds.  Now blindly tell it to exit command mode at startup...